+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## Next
 
-* Improve performance of ASCII rendering - [@dblock](https://github.com/dblock).
-* Render ASCII images on a background thread - [@dblock](https://github.com/dblock).
-* Use "Courier New" font instead of "Courier" - [@dblock](https://github.com/dblock).
+* [#2](https://github.com/dblock/ARASCIISwizzle/issues/2): Added [UIImage+ASCII](Classes/UIImage+ASCII.h) API - [@dblock](https://github.com/dblock).
+* Improved performance of ASCII rendering - [@dblock](https://github.com/dblock).
+* [#1](https://github.com/dblock/ARASCIISwizzle/issues/1): Render ASCII images on a background thread - [@dblock](https://github.com/dblock).
+* Default to the "Courier New" font instead of "Courier" - [@dblock](https://github.com/dblock).
 * Don't scale resulting images, use `UIViewContentModeScaleToFill` if you need to preserve size - [@dblock](https://github.com/dblock).
 
 ## 1.0.0 (3/21/2014)